taking out participation from within exam.js for cohesion
